# encoding/base32: increase performance and code reuse